+++ domain/domain-pos/src/main/java/domain/pos/menu/implement/MenuCategoryWriter.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Component;
 
 import domain.pos.menu.entity.MenuCategory;
 import domain.pos.menu.entity.MenuCategoryInfo;
-import domain.pos.menu.repository.MenuCategoryRepository;
+import domain.pos.menu.port.required.MenuCategoryRepository;
 import domain.pos.store.entity.Store;
 import lombok.RequiredArgsConstructor;
