+++ domain/domain-pos/src/main/java/domain/pos/menu/port/required/MenuRepository.java
@@ -1,4 +1,4 @@
-package domain.pos.menu.repository;
+package domain.pos.menu.port.required;
 
 import java.util.List;
 import java.util.Optional;
@@ -10,6 +10,7 @@ import org.springframework.stereotype.Repository;
 import domain.pos.menu.entity.Menu;
 import domain.pos.menu.entity.MenuCategoryInfo;
 import domain.pos.menu.entity.MenuInfo;
+import domain.pos.menu.entity.v2.state.CreateMenuInfoState;
 import domain.pos.store.entity.Store;
 
 @Repository
@@ -40,4 +41,18 @@ public interface MenuRepository {
 	Long countByIdIn(Long storeId, Set<Long> menuIds);
 
 	Optional<MenuInfo> getMenuInfoById(Long menuId);
+
+	// v2
+	domain.pos.menu.entity.v2.Menu createMenuAtEnd(CreateMenuInfoState createMenuInfoState, Long storeId,
+		Long menuCategoryId);
+
+	Optional<domain.pos.menu.entity.v2.Menu> readMenu(Long storeId, Long menuId);
+
+	domain.pos.menu.entity.v2.Menu updateMenuInfo(domain.pos.menu.entity.v2.Menu menu);
+
+	domain.pos.menu.entity.v2.Menu updateOrder(Long storeId, Long menuId, Integer updateOrder);
+
+	domain.pos.menu.entity.v2.Menu updateState(domain.pos.menu.entity.v2.Menu menu);
+
+	void deleteMenu(Long menuId, Long storeId);
 }
